refactor(rome_js_formatter): move language agnostic code into `rome_formatter`
