+++ crates/rome_formatter/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "rome_formatter"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rome_rowan = { path = "../rome_rowan" }
+tracing = "0.1.31"
