+++ crates/rome_formatter/src/intersperse.rs
@@ -6,7 +6,7 @@ use crate::FormatElement;
 
 /// An iterator adapter that places a separator between all elements.
 #[derive(Debug, Clone)]
-pub(crate) struct Intersperse<I: Iterator>
+pub struct Intersperse<I: Iterator>
 where
     I::Item: Clone,
 {
@@ -85,7 +85,7 @@ where
 /// The separator element is generated by a factory function called
 /// by the iterator with the previous and next SyntaxNode as arguments
 #[derive(Debug, Clone)]
-pub(crate) struct IntersperseFn<I: Iterator, F, N>
+pub struct IntersperseFn<I: Iterator, F, N>
 where
     I: Iterator<Item = (N, FormatElement)>,
 {
