+++ crates/rome_formatter/src/lib.rs
@@ -1,0 +1,168 @@
+pub mod format_element;
+pub mod format_elements;
+pub mod intersperse;
+pub mod printer;
+pub use format_element::{
+    block_indent, comment, concat_elements, empty_element, empty_line, fill_elements,
+    group_elements, hard_group_elements, hard_line_break, if_group_breaks,
+    if_group_fits_on_single_line, indent, join_elements, line_suffix, normalize_newlines,
+    soft_block_indent, soft_line_break, soft_line_break_or_space, soft_line_indent_or_space,
+    space_token, token, FormatElement, Token, Verbatim, LINE_TERMINATORS,
+};
+use rome_rowan::{TextRange, TextSize};
+use std::fmt::Display;
+use std::str::FromStr;
+
+#[derive(Debug, Eq, PartialEq, Clone, Copy)]
+pub enum IndentStyle {
+    /// Tab
+    Tab,
+    /// Space, with its quantity
+    Space(u8),
+}
+
+impl IndentStyle {
+    pub const DEFAULT_SPACES: u8 = 2;
+}
+
+impl Default for IndentStyle {
+    fn default() -> Self {
+        Self::Tab
+    }
+}
+
+impl FromStr for IndentStyle {
+    type Err = &'static str;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "tab" | "Tabs" => Ok(Self::Tab),
+            "space" | "Spaces" => Ok(Self::Space(IndentStyle::DEFAULT_SPACES)),
+            // TODO: replace this error with a diagnostic
+            _ => Err("Value not supported for IndentStyle"),
+        }
+    }
+}
+
+impl Display for IndentStyle {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            IndentStyle::Tab => write!(f, "Tab"),
+            IndentStyle::Space(size) => write!(f, "Spaces, size: {}", size),
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy)]
+pub struct FormatOptions {
+    /// The indent style
+    pub indent_style: IndentStyle,
+
+    /// What's the max width of a line. Defaults to 80
+    pub line_width: u16,
+}
+
+impl FormatOptions {
+    pub fn new(indent_style: IndentStyle) -> Self {
+        Self {
+            indent_style,
+            ..Self::default()
+        }
+    }
+}
+
+impl Default for FormatOptions {
+    fn default() -> Self {
+        Self {
+            indent_style: IndentStyle::default(),
+            line_width: 80,
+        }
+    }
+}
+
+impl Display for FormatOptions {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        writeln!(f, "Indent style: {}", self.indent_style)?;
+        writeln!(f, "Line width: {}", self.line_width)?;
+        Ok(())
+    }
+}
+
+/// Lightweight sourcemap marker between source and output tokens
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub struct SourceMarker {
+    /// Position of the marker in the original source
+    pub source: TextSize,
+    /// Position of the marker in the output code
+    pub dest: TextSize,
+}
+
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub struct Formatted {
+    code: String,
+    range: Option<TextRange>,
+    sourcemap: Vec<SourceMarker>,
+    verbatim_source: Vec<(String, TextRange)>,
+}
+
+impl Formatted {
+    pub fn new(
+        code: String,
+        range: Option<TextRange>,
+        sourcemap: Vec<SourceMarker>,
+        verbatim_source: Vec<(String, TextRange)>,
+    ) -> Self {
+        Self {
+            code,
+            range,
+            sourcemap,
+            verbatim_source,
+        }
+    }
+
+    /// Construct an empty formatter result
+    pub fn new_empty() -> Self {
+        Self {
+            code: String::new(),
+            range: None,
+            sourcemap: Vec::new(),
+            verbatim_source: Vec::new(),
+        }
+    }
+
+    /// Range of the input source file covered by this formatted code,
+    /// or None if the entire file is covered in this instance
+    pub fn range(&self) -> Option<TextRange> {
+        self.range
+    }
+
+    /// Returns a list of [SourceMarker] mapping byte positions
+    /// in the output string to the input source code
+    pub fn sourcemap(&self) -> &[SourceMarker] {
+        &self.sourcemap
+    }
+
+    /// Returns a list of [SourceMarker] mapping byte positions
+    /// in the output string to the input source code, consuming the result
+    pub fn into_sourcemap(self) -> Vec<SourceMarker> {
+        self.sourcemap
+    }
+
+    /// Access the resulting code, borrowing the result
+    pub fn as_code(&self) -> &str {
+        &self.code
+    }
+
+    /// Access the resulting code, consuming the result
+    pub fn into_code(self) -> String {
+        self.code
+    }
+
+    pub fn verbatim(&self) -> &[(String, TextRange)] {
+        &self.verbatim_source
+    }
+
+    pub fn into_verbatim(self) -> Vec<(String, TextRange)> {
+        self.verbatim_source
+    }
+}
