+++ crates/rome_js_formatter/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 [dependencies]
 rome_js_parser = { path = "../rome_js_parser" }
 rome_js_syntax = { path = "../rome_js_syntax" }
+rome_formatter = { path = "../rome_formatter" }
 rome_rowan = { path = "../rome_rowan" }
 rome_fs = { path = "../rome_fs" }
 rome_core = { path = "../rome_core" }
