+++ crates/rome_js_formatter/src/utils/mod.rs
@@ -12,6 +12,7 @@ use crate::{
 pub use binarish_expression::format_binaryish_expression;
 pub(crate) use call_expression::format_call_expression;
 pub(crate) use format_conditional::{format_conditional, Conditional};
+use rome_formatter::normalize_newlines;
 use rome_js_syntax::{AstNode, AstNodeList, JsSyntaxKind, SyntaxNode, SyntaxNodeExt, SyntaxToken};
 use rome_js_syntax::{
     JsAnyExpression, JsAnyFunction, JsAnyRoot, JsAnyStatement, JsInitializerClause,
@@ -20,7 +21,6 @@ use rome_js_syntax::{
 };
 use std::borrow::Cow;
 
-use crate::format_element::normalize_newlines;
 pub(crate) use simple::*;
 
 /// Utility function to format the separators of the nodes that belong to the unions
